Fix consolidation using incorrect fragment size
For arrays with string dimensions we must call
FragmentMetadata::get_footer_offset_and_size instead of
FragmentMetadata::get_footer_size. The get_footer_size only accounts for
fixed size fragment footers, and doesn't attempt to read the size at the
end of the footer as is required for string dimensions.